View and logic for learning mode added. Refactoring of written quiz.
