+++ EnglishWhale/View/QuizzesChooserForm.cs
@@ -48,7 +48,8 @@ namespace EnglishWhale
 
         private void learningButton_Click(object sender, EventArgs e)
         {
-
+            LanguageDictionary languageDictionary = (LanguageDictionary)diretionComboBox.SelectedItem;
+            mContr.StartLearning(languageDictionary, this);
         }
     }
 }
